Move reset password function to a separate JS file.
